Set up mod menu for now
The mod menu is kind of working for the watering can option; it's just distorted.

The math for calculating Stamina is broken. When it's called, power is "836052616" which is either the max value of the variable (unliekely), train odometer (nope), or this is an interpretation of the player's name/ID that's being sent along. The assembly shows 2 loads for the player right before calling the function which tells me it loaded the playare into both. I might need to look into how to pass the parameter to the function