Added basic settings menu
Currently there's an issue where the mod doesn't always recognize that it's time to regenerate despite the values matching. I believe this is because it's doing a less than check instead of <= so in those cases where the float is equal but actually evaluates to be less than, it may be causing issues.

Find better logic for this; I don't know if this is an issue with the original mod or not.